Shuffle some tests around
