+++ python/tests/unit/arcticdb/version_store/test_append.py
@@ -2,23 +2,15 @@ import random
 import pandas as pd
 import numpy as np
 from datetime import datetime
-from itertools import chain, product, combinations
 import pytest
 import sys
-import os
 from numpy.testing import assert_array_equal
-from hypothesis import given, assume, settings, strategies as st
 
 from pandas import MultiIndex
-from arcticdb.version_store._common import TimeFrame
 from arcticdb.version_store import NativeVersionStore
-from arcticdb.util.test import random_integers, assert_frame_equal, random_seed_context
-from arcticdb.util.hypothesis import InputFactories, use_of_function_scoped_fixtures_in_hypothesis_checked
 from arcticdb_ext.exceptions import InternalException, NormalizationException, SortingException
 from arcticdb_ext import set_config_int
-
-from tests.util.mark import SLOW_TESTS_MARK
-
+from arcticdb.util.test import random_integers, assert_frame_equal
 
 def test_append_simple(lmdb_version_store):
     symbol = "test_append_simple"
@@ -64,88 +56,6 @@ def test_append_string_of_different_sizes(lmdb_version_store):
     vit = lmdb_version_store.read(symbol)
     expected = pd.concat([df1, df2])
     assert_frame_equal(vit.data, expected)
-
-
-def gen_params_append():
-    with random_seed_context():
-        # colnums
-        p = [list(range(2, 5))]
-        # periods
-        periods = 6
-        p.append([periods])
-        # rownums
-        p.append([1, 4, periods + 2])
-        # cols
-        p.append(list(chain(*[list(combinations(["a", "b", "c"], c)) for c in range(1, 4, 2)])))
-        # tsbounds
-        p.append([(j, i) for i in [1, periods - 1] for j in range(i)])
-        # append_point
-        p.append([k for k in range(1, periods - 1)])
-        result = sorted(random.sample(list(product(*p)), 10))
-    return result
-
-
-def gen_params_append_single():
-    # colnums
-    p = [[2]]
-    # periods
-    periods = 6
-    p.append([periods])
-    # rownums
-    p.append([1])
-    # cols
-    p.append([["a"]])
-    # tsbounds
-    p.append([[0, 5]])
-    # append_point
-    p.append([1])
-    return list(product(*p))
-
-
-@pytest.mark.parametrize("colnum,periods,rownum,cols,tsbounds,append_point", gen_params_append())
-def test_append_partial_read(version_store_factory, colnum, periods, rownum, cols, tsbounds, append_point):
-    tz = "America/New_York"
-    version_store = version_store_factory(col_per_group=colnum, row_per_segment=rownum)
-    dtidx = pd.date_range("2019-02-06 11:43", periods=6).tz_localize(tz)
-    a = np.arange(dtidx.shape[0])
-    tf = TimeFrame(dtidx.values, columns_names=["a", "b", "c"], columns_values=[a, a + a, a * 10])
-    c1 = dtidx[append_point]
-    c2 = dtidx[append_point + 1]
-    tf1 = tf.tsloc[:c1]
-    sid = "XXX"
-    version_store.write(sid, tf1)
-    tf2 = tf.tsloc[c2:]
-    version_store.append(sid, tf2)
-
-    dtr = (dtidx[tsbounds[0]], dtidx[tsbounds[1]])
-    vit = version_store.read(sid, date_range=dtr, columns=list(cols))
-    rtf = tf.tsloc[dtr[0] : dtr[1]]
-    col_names, col_values = zip(*[(c, v) for c, v in zip(rtf.columns_names, rtf.columns_values) if c in cols])
-    rtf = TimeFrame(rtf.times, list(col_names), list(col_values))
-    assert rtf == vit.data
-
-
-@pytest.mark.parametrize("colnum,periods,rownum,cols,tsbounds,append_point", gen_params_append())
-def test_incomplete_append_partial_read(version_store_factory, colnum, periods, rownum, cols, tsbounds, append_point):
-    tz = "America/New_York"
-    version_store = version_store_factory(col_per_group=colnum, row_per_segment=rownum)
-    dtidx = pd.date_range("2019-02-06 11:43", periods=6).tz_localize(tz)
-    a = np.arange(dtidx.shape[0])
-    tf = TimeFrame(dtidx.values, columns_names=["a", "b", "c"], columns_values=[a, a + a, a * 10])
-    c1 = dtidx[append_point]
-    c2 = dtidx[append_point + 1]
-    tf1 = tf.tsloc[:c1]
-    sid = "XXX"
-    version_store.write(sid, tf1)
-    tf2 = tf.tsloc[c2:]
-    version_store.append(sid, tf2, incomplete=True)
-
-    dtr = (dtidx[tsbounds[0]], dtidx[tsbounds[1]])
-    vit = version_store.read(sid, date_range=dtr, columns=list(cols), incomplete=True)
-    rtf = tf.tsloc[dtr[0] : dtr[1]]
-    col_names, col_values = zip(*[(c, v) for c, v in zip(rtf.columns_names, rtf.columns_values) if c in cols])
-    rtf = TimeFrame(rtf.times, list(col_names), list(col_values))
-    assert rtf == vit.data
 
 
 def test_append_snapshot_delete(lmdb_version_store):
@@ -265,31 +175,6 @@ def test_append_pickled_symbol(lmdb_version_store):
     assert lmdb_version_store.is_symbol_pickled(symbol)
     with pytest.raises(InternalException):
         _ = lmdb_version_store.append(symbol, np.arange(100).tolist())
-
-
-# test_hypothesis_version_store.py covers the appends that are allowed.
-# Only need to check the forbidden ones have useful error messages:
-@pytest.mark.parametrize(
-    "initial, append, match",
-    [
-        # (InputFactories.DF_RC_NON_RANGE, InputFactories.DF_DTI, "TODO(AN-722)"),
-        (InputFactories.DF_RC, InputFactories.ND_ARRAY_1D, "(Pandas|ndarray)"),
-        (InputFactories.DF_RC, InputFactories.DF_MULTI_RC, "index type incompatible"),
-        (InputFactories.DF_RC, InputFactories.DF_RC_NON_RANGE, "range.*index which is incompatible"),
-        (InputFactories.DF_RC, InputFactories.DF_RC_STEP, "different.*step"),
-    ],
-)
-@pytest.mark.parametrize("swap", ["swap", ""])
-def test_(initial: InputFactories, append: InputFactories, match, swap, lmdb_version_store: NativeVersionStore):
-    lib = lmdb_version_store
-    if swap:
-        initial, append = append, initial
-    init_data, next_start = initial.make(1, 3)
-    lib.write("s", init_data)
-
-    to_append, _ = append.make(abs(next_start), 1)
-    with pytest.raises(NormalizationException):
-        lib.append("s", to_append, match=match)
 
 
 def test_append_not_sorted_exception(lmdb_version_store):
@@ -524,134 +409,6 @@ def test_append_mix_ascending_descending(lmdb_version_store):
     lmdb_version_store.append(symbol, df2)
     info = lmdb_version_store.get_info(symbol)
     assert info["sorted"] == "UNSORTED"
-
-
-@use_of_function_scoped_fixtures_in_hypothesis_checked
-@settings(deadline=None)
-@given(
-    col_per_append_df=st.integers(2, 100),
-    col_name_set=st.integers(1, 10000),
-    num_rows_per_test_cycle=st.lists(st.lists(st.integers(1, 20), min_size=1, max_size=10), max_size=2),
-    column_group_size=st.integers(2, 100),
-    segment_row_size=st.integers(2, 100),
-    dynamic_schema=st.booleans(),
-    dynamic_strings=st.booleans(),
-    df_in_str=st.booleans(),
-)
-@SLOW_TESTS_MARK
-def test_append_with_defragmentation(
-    sym,
-    col_per_append_df,
-    col_name_set,
-    num_rows_per_test_cycle,
-    get_wide_df,
-    column_group_size,
-    segment_row_size,
-    dynamic_schema,
-    dynamic_strings,
-    df_in_str,
-    basic_store_factory,
-):
-    def get_wide_and_long_df(start_idx, end_idx, col_per_append_df, col_name_set, df_in_str):
-        df = pd.DataFrame()
-        for idx in range(start_idx, end_idx):
-            df = pd.concat([df, get_wide_df(idx, col_per_append_df, col_name_set)])
-        if col_per_append_df == col_name_set:  # manually sort them for static schema, for newer version of panda
-            df = df.reindex(sorted(list(df.columns)), axis=1)
-        df = df.astype(str if df_in_str else np.float64)
-        return df
-
-    def get_no_of_segments_after_defragmentation(df, merged_segment_row_size):
-        new_segment_row_size = no_of_segments = 0
-        for start_row, end_row in pd.Series(df.end_row.values, index=df.start_row).to_dict().items():
-            no_of_segments = no_of_segments + 1 if new_segment_row_size == 0 else no_of_segments
-            new_segment_row_size += end_row - start_row
-            if new_segment_row_size >= merged_segment_row_size:
-                new_segment_row_size = 0
-        return no_of_segments
-
-    def get_no_of_column_merged_segments(df):
-        return len(pd.Series(df.end_row.values, index=df.start_row).to_dict().items())
-
-    def run_test(
-        lib,
-        col_per_append_df,
-        col_name_set,
-        merged_segment_row_size,
-        df_in_str,
-        before_compact,
-        index_offset,
-        num_of_rows,
-    ):
-        for num_of_row in num_of_rows:
-            start_index = index_offset
-            index_offset += num_of_row
-            end_index = index_offset
-            df = get_wide_and_long_df(start_index, end_index, col_per_append_df, col_name_set, df_in_str)
-            before_compact = pd.concat([before_compact, df])
-            if start_index == 0:
-                lib.write(sym, df)
-            else:
-                lib.append(sym, df)
-            segment_details = lib.read_index(sym)
-            assert lib.is_symbol_fragmented(sym, None) is (
-                get_no_of_segments_after_defragmentation(segment_details, merged_segment_row_size)
-                != get_no_of_column_merged_segments(segment_details)
-            )
-        if get_no_of_segments_after_defragmentation(
-            segment_details, merged_segment_row_size
-        ) != get_no_of_column_merged_segments(segment_details):
-            seg_details_before_compaction = lib.read_index(sym)
-            lib.defragment_symbol_data(sym, None)
-            res = lib.read(sym).data
-            res = res.reindex(sorted(list(res.columns)), axis=1)
-            res = res.replace("", 0.0)
-            res = res.fillna(0.0)
-            before_compact = before_compact.reindex(sorted(list(before_compact.columns)), axis=1)
-            before_compact = before_compact.fillna(0.0)
-
-            seg_details = lib.read_index(sym)
-
-            assert_frame_equal(before_compact, res)
-
-            assert len(seg_details) == get_no_of_segments_after_defragmentation(
-                seg_details_before_compaction, merged_segment_row_size
-            )
-            indexs = (
-                seg_details["end_index"].astype(str).str.rsplit(" ", n=2).agg(" ".join).reset_index()
-            )  # start_index and end_index got merged into one column
-            assert np.array_equal(indexs.iloc[1:, 0].astype(str).values, indexs.iloc[:-1, 1].astype(str).values)
-        else:
-            with pytest.raises(InternalException):
-                lib.defragment_symbol_data(sym, None)
-        return before_compact, index_offset
-
-    assume(col_per_append_df <= col_name_set)
-    assume(
-        num_of_row % 2 != 0 for num_of_rows in num_rows_per_test_cycle for num_of_row in num_of_rows
-    )  # Make sure at least one successful compaction run per cycle
-
-    set_config_int("SymbolDataCompact.SegmentCount", 1)
-    before_compact = pd.DataFrame()
-    index_offset = 0
-    lib = basic_store_factory(
-        column_group_size=column_group_size,
-        segment_row_size=segment_row_size,
-        dynamic_schema=dynamic_schema,
-        dynamic_strings=dynamic_strings,
-        reuse_name=True,
-    )
-    for num_of_rows in num_rows_per_test_cycle:
-        before_compact, index_offset = run_test(
-            lib,
-            col_per_append_df,
-            col_name_set if dynamic_schema else col_per_append_df,
-            segment_row_size,
-            df_in_str,
-            before_compact,
-            index_offset,
-            num_of_rows,
-        )
 
 
 def test_append_with_cont_mem_problem(sym, lmdb_version_store_tiny_segment_dynamic):
